pull out util methods into separate class for easy removal
